refactor(client): Refactor helper code into its own file
